pyproject.toml
[project]
name = "snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "strands-agents[anthropic]>=1.22.0",
    "strands-agents-tools>=0.2.19",
    "requests>=2.32.5",
    "markdown>=3.10.1",
    "python-dotenv>=1.0.0",
]
